+++ src/components/Auth/UserAuth/UserAuth.module.scss
@@ -24,8 +24,7 @@
     border-radius: 0.4rem;
     padding: 0 2.4rem;
 
-    @include headerLink-hover();
-    letter-spacing: calc(1.2rem * 0.09);
+    @include headerLink-hover(9);
 
     &.SigninBtn {
       background-color: transparent;
