+++ src/scss/1-Helpers/_mixin.scss
@@ -102,45 +102,14 @@
   }
 }
 
-@mixin screen($size, $pixels: brk('sm')) {
-  @if $size == 'sm' {
-    @media screen and (min-width: brk('sm')) {
+@mixin screen($name) {
+  @if map-has-key($map: $breakpoints, $key: $name) {
+    @media #{inspect(map-get($breakpoints, $name))} {
       @content;
     }
-  } @else if $size == 'md' {
-    @media screen and (min-width: brk('md')) {
-      @content;
-    }
-  } @else if $size == 'lg' {
-    @media screen and (min-width: brk('lg')) {
-      @content;
-    }
-  } @else if $size == 'x-lg' {
-    @media screen and (min-width: brk('x-lg')) {
-      @content;
-    }
-  } @else if $size == 'custom' {
-    @media screen and (min-width: $pixels + px) {
-      @content;
-    }
-  }
-}
-
-@mixin headerLink-hover() {
-  color: $txt-white;
-  font-size: 16px;
-
-  &:hover,
-  &:active,
-  &:visited {
-    color: $txt-headerLink-hover;
-  }
-
-  @include screen(lg) {
-    color: $txt-headerLink;
-    text-transform: uppercase;
-    font-weight: bold;
-    font-size: 12px;
+  } @else {
+    @warn "Unfortunately, no value could be retrieved from `#{$breakpoint}`. "
+    + "Please make sure it is defined in `$breakpoints` map.";
   }
 }
