+++ src/scss/3-Layout/_header.scss
@@ -1,0 +1,16 @@
+@mixin headerLink-hover($letter-spacing: 0) {
+  @include font-size($fs-headerLink, $letter-spacing);
+  color: $txt-white;
+
+  &:hover,
+  &:active,
+  &:visited {
+    color: $txt-headerLink-hover;
+  }
+
+  @include screen(lg) {
+    color: $txt-headerLink;
+    text-transform: uppercase;
+    font-weight: bold;
+  }
+}
