+++ src/scss/_main.scss
@@ -2,3 +2,5 @@
 @import './1-Helpers/function';
 @import './1-Helpers/mixin';
 @import './1-Helpers/variable';
+
+@import './3-Layout/header';
